move action menu tests to separate file
